added stuff for it to run on battlegrounds server

+++ cogs/on_message.py
@@ -30,13 +30,14 @@ class On_message(commands.Cog):
         #searches for invite links, no need for regex as there are a million ways to disguise invite links anyways. 
         #this is just a basic filter which is not designed to stop 100% of invite links coming in
         if "discord.gg" in message.content or "discordapp.com/invite" in message.content:
-            guild = self.bot.get_guild(739299507795132486)
-            mod_role = discord.utils.get(guild.roles, id=739299507816366106)
-            promoter_role = discord.utils.get(guild.roles, id=739299507799326847)
-            allowed_channels = (739299509670248502, 739299508902559811, 739299508403437623, 739299508197917060, 739299509670248505)
-            if mod_role not in message.author.roles and promoter_role not in message.author.roles and message.channel.id not in allowed_channels:
-                await message.delete()
-                await message.channel.send(f"Please don't send invite links here {message.author.mention}")
+            if message.guild.id == 739299507795132486:
+                guild = self.bot.get_guild(739299507795132486)
+                mod_role = discord.utils.get(guild.roles, id=739299507816366106)
+                promoter_role = discord.utils.get(guild.roles, id=739299507799326847)
+                allowed_channels = (739299509670248502, 739299508902559811, 739299508403437623, 739299508197917060, 739299509670248505)
+                if mod_role not in message.author.roles and promoter_role not in message.author.roles and message.channel.id not in allowed_channels:
+                    await message.delete()
+                    await message.channel.send(f"Please don't send invite links here {message.author.mention}")
 
 
         #searches for bad words using regex
